MR-19 added food plan screen and data, domain layer

+++ app/src/main/java/com/example/myration/ui/FoodPlanScreen/ManageFoodPlanScreen.kt
@@ -173,7 +173,7 @@ fun ApprovePlanDialogue(
                                 horizontalArrangement = Arrangement.SpaceBetween
                             ) {
                                 Text(foodPlans[index].mealName)
-                                Text(stringResource(R.string.meal_calorie_kcal, foodPlans[index].mealCalorie), color = Color.Gray)
+                                Text(stringResource(R.string.meal_calorie_kcal, foodPlans[index].mealCalorie.toInt()), color = Color.Gray)
                             }
                         }
                     )
